Add files via upload
Homelessness in the USA is one of the significant problems associated with the nation’s current-day development.
The data we got from the United States census bureau has helped us visualize in this hackathon.
Apart from this, the information we use in this dataset has been taken from the national alliance to end
homelessness, which allowed us to analyze the different types of homelessness and relate homelessness to
the chronic diseases people are suffering from. The Homeless research Institute has been trying to reduce
homelessness over the past few years.


Problem Statement
Working on these datasets based on 7-8 years old helped us read the scenario of the dataset. We have seen
all the different possibilities of homelessness and the reasons for homelessness. Most chronically diseased
people are homeless; followed by this, the dataset significantly impacts individuals rather than the families.
In addition to this, major cities have more homeless people like New York, Massachusetts, North Dakota,
Oregon, Nevada, and California. Initially, we established the trends in the dataset by following the data
through the years 2005-2013. Supplementing this, we found out the different categories of homeless people
in the year 2013. We split the data into two categories based on population to establish a relationship to
look into the change over last year’s homeless people count. Furthermore, the homeless population also had
non-chronic who were lost, which helped us find the difference between chronic and non-chronic people who
were homeless. This data was until 2013, but as the trends keep on updating, the data keeps on updating,
helping us analyze the decrement of homeless over the years.

Help in accessing files -
1. The pdf file is the complete summary of the entire project with visualizations and analysis.
2. The .Rmd file contains the R code for the analysis and visualizations